cast indices.size() to int

+++ OpenGl Game/OpenGlGame/resources/mesh.h
@@ -41,9 +41,10 @@ namespace glg {
 			}
 			
 			glBindVertexArray(VAO);
-			glDrawElements(GL_TRIANGLES, indices.size(), GL_UNSIGNED_INT, 0);
+			glDrawElements(GL_TRIANGLES, (int) indices.size(), GL_UNSIGNED_INT, 0);
 			glBindVertexArray(0);
 			glActiveTexture(GL_TEXTURE0);
+
 		}
 
 	private:
